fix: address comments
